Refactor Crashlytics log & key management classes into new `metadata` package.
