+++ firebase-crashlytics/src/main/java/com/google/firebase/crashlytics/internal/metadata/FileLogStore.java
@@ -12,5 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/** @hide */
-package com.google.firebase.crashlytics.internal.log;
+package com.google.firebase.crashlytics.internal.metadata;
+
+import androidx.annotation.Nullable;
+
+interface FileLogStore {
+  void writeToLog(long timestamp, String msg);
+
+  /** Get log file contents as bytes. Returns null if no data is available. */
+  @Nullable
+  byte[] getLogAsBytes();
+
+  /** Get log file contents as a String. Returns null if no data is available. */
+  @Nullable
+  String getLogAsString();
+
+  void closeLogFile();
+
+  void deleteLogFile();
+}
