+++ firebase-crashlytics/src/main/java/com/google/firebase/crashlytics/internal/metadata/QueueFile.java
@@ -19,7 +19,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.firebase.crashlytics.internal.log;
+package com.google.firebase.crashlytics.internal.metadata;
 
 import java.io.Closeable;
 import java.io.File;
