+++ firebase-crashlytics/src/main/java/com/google/firebase/crashlytics/internal/metadata/package-info.java
@@ -12,22 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.crashlytics.internal.log;
-
-import androidx.annotation.Nullable;
-
-interface FileLogStore {
-  void writeToLog(long timestamp, String msg);
-
-  /** Get log file contents as bytes. Returns null if no data is available. */
-  @Nullable
-  byte[] getLogAsBytes();
-
-  /** Get log file contents as a String. Returns null if no data is available. */
-  @Nullable
-  String getLogAsString();
-
-  void closeLogFile();
-
-  void deleteLogFile();
-}
+/** @hide */
+package com.google.firebase.crashlytics.internal.metadata;
